requirements.txt
vllm
fastapi
uvicorn
python-dotenv
langchain
tiktoken
pydantic
skyfield
geopy
